Rename 'backend' module to 'database'
