+++ views/groups.go
@@ -2,7 +2,7 @@ package views
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/ordinary-dev/phoenix/backend"
+	"github.com/ordinary-dev/phoenix/database"
 	"gorm.io/gorm"
 	"net/http"
 	"strconv"
@@ -10,7 +10,7 @@ import (
 
 func CreateGroup(c *gin.Context, db *gorm.DB) {
 	// Save new group to the database.
-	group := backend.Group{
+	group := database.Group{
 		Name: c.PostForm("groupName"),
 	}
 	if result := db.Create(&group); result.Error != nil {
@@ -29,7 +29,7 @@ func UpdateGroup(c *gin.Context, db *gorm.DB) {
 		return
 	}
 
-	var group backend.Group
+	var group database.Group
 	if result := db.First(&group, id); result.Error != nil {
 		ShowError(c, result.Error)
 		return
@@ -52,7 +52,7 @@ func DeleteGroup(c *gin.Context, db *gorm.DB) {
 		return
 	}
 
-	if result := db.Delete(&backend.Group{}, id); result.Error != nil {
+	if result := db.Delete(&database.Group{}, id); result.Error != nil {
 		ShowError(c, result.Error)
 		return
 	}
